Replace ssize_t with long, as Win32 lacks proper support for that type.
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@47197 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/session/mod_files.c
@@ -246,7 +246,7 @@ PS_CLOSE_FUNC(files)
 
 PS_READ_FUNC(files)
 {
-	ssize_t n;
+	long n;
 	struct stat sbuf;
 	PS_FILES_DATA;
 
@@ -276,7 +276,7 @@ PS_READ_FUNC(files)
 
 PS_WRITE_FUNC(files)
 {
-	ssize_t n;
+	long n;
 	PS_FILES_DATA;
 
 	ps_files_open(data, key);
